refactor: keep HTTP redirect policy in portal
